backend/requirements.txt
fastapi
uvicorn[gunicorn]
python-multipart
ffmpeg-python


